Rename plots module
