+++ simulation/elevator/simulator.py
@@ -5,7 +5,6 @@ import numpy as np
 from typing import List
 from dataclasses import asdict, dataclass, field
 
-from . import plots
 from .runtime import Config
